fixed re-renders in TodoContainer. Refactored index, app. Created approutes.js
