+++ src/App.js
@@ -1,20 +1,29 @@
-import React from 'react';
-import TodoContainer from './components/TodoContainer';
-import { BrowserRouter, Route, Routes } from 'react-router-dom';
-import './components/TodoListItem.module.css';
+import React, { useState } from 'react';
+import AppRoutes from './AppRoutes';
+import SideDiv from './components/SideDiv';
+import style from './components/TodoListItem.module.css';
 
-function App({ selectedList }) {
+function App() {
+  const listNames = ['Work', 'Personal', 'School'];
+  const [selectedList, setSelectedList] = useState('Personal');
+  const onSelectList = (listName) => {
+    // Handle the selected list as needed
+    console.log(`Selected List: ${listName}`);
+
+    // Update the selectedList state
+    setSelectedList(listName);
+  };
+
+
   return (
-    <BrowserRouter>
-      <Routes>
-        <Route path="/" element={<TodoContainer selectedList={selectedList} />} />
-        <Route path="/new" element={
-          <div style={{ textAlign: 'center' }}>
-            <h1>New Todo List</h1>
-          </div>
-        } />
-      </Routes>
-    </BrowserRouter>
+    <div className={style.container}>
+      <div className={style.box}>
+        <SideDiv listNames={listNames} onSelectList={onSelectList} />
+      </div>
+      <div className={style.box} >
+        <AppRoutes selectedList={selectedList} />
+      </div>
+    </div>
   );
 }
